Added Matrices Exercises (2/12) & Exam Prep 3 (3/4)

+++ C# Advanced/03. Matrices/Lab/04. Pascal Triangle/Program.cs
@@ -6,26 +6,26 @@ namespace _04.Pascal_Triangle
     {
         static void Main(string[] args)
         {
-	        int n = int.Parse(Console.ReadLine());
-			int[][] matrix = new int[n][];
+	        long n = long.Parse(Console.ReadLine());
+			long[][] matrix = new long[n][];
 
-			matrix[0] = new int[1];
+			matrix[0] = new long[1];
 	        matrix[0][0] = 1;
 
-	        for (int i = 1; i < n; i++)
+	        for (long i = 1; i < n; i++)
 	        {
-				matrix[i] = new int[i + 1];
+				matrix[i] = new long[i + 1];
 		        matrix[i][0] = 1;
 		        matrix[i][i] = 1;
 
-		        for (int j = 1; j < i; j++)
+		        for (long j = 1; j < i; j++)
 		        {
 			        matrix[i][j] = matrix[i - 1][j] + matrix[i - 1][j - 1];
 		        }
 	        }
 
 	        string result = "";
-	        foreach (int[] row in matrix)
+	        foreach (long[] row in matrix)
 	        {
 		        result += $"{string.Join(" ", row)}\n";
 	        }
